+++ FoodRecipes/app/build.gradle
@@ -39,6 +39,7 @@ dependencies {
     def glide_version = "4.11.0"
     def coordinate_layout_version = "1.1.0"
     def material_version = "1.0.0"
+    def circle_image_view_version = "3.1.0"
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
@@ -74,4 +75,6 @@ dependencies {
     implementation "androidx.coordinatorlayout:coordinatorlayout:$coordinate_layout_version"
     //Material
     implementation "com.google.android.material:material:$material_version"
+    //Circle Image View
+    implementation "de.hdodenhof:circleimageview:$circle_image_view_version"
 }
