+++ FoodRecipes/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.chrislicoder.foorecipes">
+    package="com.chrislicoder.foodrecipes">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -11,7 +11,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".RecipeListActivity">
+        <activity android:name="com.chrislicoder.foodrecipes.RecipeListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
