+++ FoodRecipes/app/src/main/java/com/chrislicoder/foodrecipes/adapters/OnRecipeListener.kt
@@ -1,0 +1,6 @@
+package com.chrislicoder.foodrecipes.adapters
+
+interface OnRecipeListener {
+    fun onRecipeClick(position: Int)
+    fun onCategoryClick(category: String)
+}
