+++ FoodRecipes/app/src/main/java/com/chrislicoder/foodrecipes/adapters/viewholders/LoadingViewHolder.kt
@@ -1,0 +1,7 @@
+package com.chrislicoder.foodrecipes.adapters.viewholders
+
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class LoadingViewHolder(itemView: View) :
+    RecyclerView.ViewHolder(itemView)
