+++ FoodRecipes/app/src/main/java/com/chrislicoder/foodrecipes/requests/ServiceGenerator.kt
@@ -1,0 +1,15 @@
+package com.chrislicoder.foodrecipes.requests
+
+import com.chrislicoder.foodrecipes.util.Constants.BASE_URL
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object ServiceGenerator {
+    val recipeApi: RecipeApi by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+            .create(RecipeApi::class.java)
+    }
+}
