+++ FoodRecipes/app/src/main/java/com/chrislicoder/foodrecipes/util/Constants.kt
@@ -1,0 +1,29 @@
+package com.chrislicoder.foodrecipes.util
+
+object Constants {
+    const val BASE_URL = "https://recipesapi.herokuapp.com"
+    const val API_KEY = ""
+    const val NETWORK_TIMEOUT = 3000
+
+    val DEFAULT_SEARCH_CATEGORIES = arrayOf(
+        "Barbeque",
+        "Breakfast",
+        "Chicken",
+        "Beef",
+        "Brunch",
+        "Dinner",
+        "Wine",
+        "Italian"
+    )
+
+    val DEFAULT_SEARCH_CATEGORY_IMAGES = arrayOf(
+        "barbeque",
+        "breakfast",
+        "chicken",
+        "beef",
+        "brunch",
+        "dinner",
+        "wine",
+        "italian"
+    )
+}
