--- FoodRecipes/app/src/main/java/com/chrislicoder/foorecipes/adapters/LoadingViewHolder.kt
@@ -1,7 +1,0 @@
-package com.chrislicoder.foorecipes.adapters
-
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-class LoadingViewHolder(itemView: View) :
-    RecyclerView.ViewHolder(itemView)
