--- FoodRecipes/app/src/main/java/com/chrislicoder/foorecipes/adapters/OnRecipeListener.kt
@@ -1,6 +1,0 @@
-package com.chrislicoder.foorecipes.adapters
-
-interface OnRecipeListener {
-    fun onRecipeClick(position: Int)
-    fun onCategoryClick(category: String)
-}
