--- FoodRecipes/app/src/main/java/com/chrislicoder/foorecipes/requests/ServiceGenerator.kt
@@ -1,15 +1,0 @@
-package com.chrislicoder.foorecipes.requests
-
-import com.chrislicoder.foorecipes.util.Constants.BASE_URL
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object ServiceGenerator {
-    val recipeApi: RecipeApi by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(RecipeApi::class.java)
-    }
-}
