--- FoodRecipes/app/src/main/java/com/chrislicoder/foorecipes/util/Constants.kt
@@ -1,7 +1,0 @@
-package com.chrislicoder.foorecipes.util
-
-object Constants {
-    const val BASE_URL = "https://recipesapi.herokuapp.com"
-    const val API_KEY = ""
-    const val NETWORK_TIMEOUT = 3000
-}
