+++ FoodRecipes/app/src/main/res/layout/layout_loading_list_item.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="@dimen/recycler_loading_view_height">
 
-    <com.chrislicoder.foorecipes.util.UI.HorizontalDottedProgress
+    <com.chrislicoder.foodrecipes.util.UI.HorizontalDottedProgress
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_centerInParent="true" />
